sdk_builder/frontmatter_builder/requirements.txt
openhands-sdk>=0.1.0
pydantic>=2.0.0